revert to ants transforms
Did comparisons between ants and hcp provided transforms on discrete subcortical
segmentations. Qualitatively, the ants transformations appeared to be
better. Reverting back for now, but can make change back to hcp provided
in the future. Additional thought was that the ants provided
transformations were more generalizable in the future as the workflow is
already available to generate new transforms.

Also:
  - Reimplemented use of parallel
  - Updated .gitignore
  - Simplify inputs and outputs of hcp_mmp.smk