Add singularity call to fslmaths
In the rule `track_from_clusters`, `fslmaths` is called directly. This only works if `fslmaths` is installed on the system, otherwise it ambiguously errors out. 

This commit instead calls `fslmaths` from the `fsl_cuda` container that is being used for `probtrackx2_gpu` (adding `singularity exec -e {params.container}`prior to the `fslmaths` call).